fix length to int from double

+++ lib/models/models.dart
@@ -1,7 +1,7 @@
 class AudioCategory {
   final String? coverImage;
   final String? name;
-  final double? length;
+  final int? length;
   final String? category;
 
   AudioCategory(this.coverImage, this.name, this.length, this.category);
